Bugfix: make "safe" char offset actually safe
The debugger crashed when stopping at a statement that included a blank line. Golfed test case:

{1

} 

I did not see any new problems when testing this bugfix with the test case, some variants of it, and with a 200 line script that I'm working on and which happened to trigger this bug.