Speedup display of regex match
Only subst in the couple hundred chars around the current match position.

If the target is hundreds or thousands of lines long then this is always going to be a lot faster. Right now the impact is colossal for target strings of more than a dozen typical text file lines.